pyproject.toml
[project]
name = "audiospectr"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "librosa>=0.10",
    "scipy>=1.11",
    "soundfile>=0.12",
    "pydub>=0.25",
    "pillow>=10.0",
    "typer>=0.9",
    "rich>=13.0",
    "pydantic>=2.0",
]

[project.scripts]
spectr = "app.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
artifacts = ["presets"]

[dependency-groups]
dev = [
    "pytest>=7.4",
    "pytest-mock>=3.12",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
